Logfile playback: Only read one packet at a time
This is a less efficient approach to parsing the logfile, since lots of
small file operations is slower than one large file operation. However, it
doesn't make much of a performance difference so far, and it saves
significant development time.

A good time to "fix" this inefficiency would be when adding support for
random playback access.