UI: Add timestamp data
This is a first attempt to show the user when the packet was received.
There also is color highlighting of atomic packets in order to link the
timestamp with the packet itself.

NOTE: Certain text append operations can be very slow. This code was
optimized using data in https://bugreports.qt-project.org/browse/QTBUG-3554